tests/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
flask==3.0.0
pandas==2.0.3
numpy==1.24.3
beautifulsoup4==4.12.2
selenium==4.10.0
tweepy==4.14.0
openai==0.27.8
pillow==10.0.0
moviepy==1.0.3
yt-dlp==2023.7.6
aiohttp==3.8.5
sqlalchemy==2.0.19
pydantic==2.0.2
python-telegram-bot==13.14
schedule==1.2.0
colorama==0.4.6
tqdm==4.65.0
psycopg2-binary==2.9.6
redis==4.5.5
celery==5.3.1